% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_unobserved.R
\name{impute_unobserved}
\alias{impute_unobserved}
\title{Impute Unobserved Potential Outcomes}
\usage{
impute_unobserved(schedule, tau)
}
\arguments{
\item{schedule}{A data frame containing the schedule of potential outcomes, with one row per unit and one column for each treatment level.}

\item{tau}{A numeric value of the Individual Treatment Effect to use to impute for the unobserved potential outcomes.}
}
\value{
A data frame with the same structure as \code{schedule} but with the unobserved potential outcomes imputed using the specified value of \code{tau}.
}
\description{
Impute the unobserved potential outcomes in a schedule of potential outcomes using a specified value of the Individual Treatment Effect (ITE).
}
\examples{
#' data <- data.frame(
 id = 1:6,
 treatment = factor(c("control", "control", "treatment", "treatment", "control", "treatment")),
 response = c(10, 12, 15, 18, 11, 17)
)
schedule <- as_schedule(data, treatment, response)

# the fisher sharp null: tau = 0
imputed_schedule_0 <- impute_unobserved(schedule, tau = 0)
print(imputed_schedule_0)

# a constant shift effect size of 3
imputed_schedule_3 <- impute_unobserved(schedule, tau = 3)
print(imputed_schedule_3)
}
