% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule.R
\name{as_schedule}
\alias{as_schedule}
\title{Convert data frame to a schedule of potential outcomes}
\usage{
as_schedule(data, id, treatment, response, append = TRUE)
}
\arguments{
\item{data}{A data frame of data from a randomized experiment.}

\item{id}{A character string of the name of the column in \code{data} that contains the unit identifiers.}

\item{treatment}{A character string of the name of the column in \code{data} that contains the treatment assignments.}

\item{response}{A character vector of the name of the column in \code{data} that contains the response variable.}

\item{append}{A logical value indicating whether to append the schedule of potential outcomes to the original data frame.}
}
\value{
The original data frame with the schedule of potential outcomes appended as new columns.
}
\description{
Convert data frame to a schedule of potential outcomes
}
