% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_schedule.R
\name{as_schedule}
\alias{as_schedule}
\title{Convert data frame to a schedule of potential outcomes}
\usage{
as_schedule(data, treatment, response, append = TRUE, outcome_prefix = "Y")
}
\arguments{
\item{data}{A data frame of data from a randomized experiment.}

\item{treatment}{Name of the column in \code{data} that contains the treatment assignments, a factor.}

\item{response}{Name of the column in \code{data} that contains the response variable.}

\item{append}{A logical value indicating whether to append the schedule of potential outcomes to the original data frame.}

\item{outcome_prefix}{A character string to prefix the names of the columns containing the potential outcomes.}
}
\value{
A data frame with the schedule of potential outcomes appended as new columns.
}
\description{
Convert data frame to a schedule of potential outcomes
}
\examples{
data <- data.frame(
 id = 1:6,
 treatment = factor(c("control", "control", "treatment", "treatment", "control", "treatment")),
 response = c(10, 12, 15, 18, 11, 17)
)
schedule <- as_schedule(data, treatment, response)
print(schedule)
}
