% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_unobserved.R
\name{impute_unobserved}
\alias{impute_unobserved}
\title{Impute Unobserved Potential Outcomes}
\usage{
impute_unobserved(schedule, tau)
}
\arguments{
\item{schedule}{A data frame containing the schedule of potential outcomes, with one row per unit and one column for each treatment level.}

\item{tau}{A numeric value of the Individual Treatment Effect to use to impute for the unobserved potential outcomes.}
}
\value{
A data frame with the same structure as \code{schedule} but with the unobserved potential outcomes imputed using the specified value of \code{tau}.
}
\description{
Impute the unobserved potential outcomes in a schedule of potential outcomes using a specified value of the Individual Treatment Effect (ITE).
}
